Initial commit.
Put together a blend file to have reasonable defaults for generating
rings, but ideally the only thing that will need to be checked in going
forwards is just the python script.

Python script generates two rings, one for me, and one for Mym.  Sizing
is pulled out of a dictionary.  Still trying to find a good base size
for both of us, then we can get fancy.